moved plot methods to diagnostics and updated README
